+++ src/ZServer/FTPServer.py
@@ -64,6 +64,7 @@ FTP Authorization
 
 """
 
+import sys
 from PubCore import handle
 from medusa.ftp_server import ftp_channel, ftp_server, recv_channel
 import asyncore, asynchat
@@ -80,6 +81,8 @@ from mimetypes import guess_type
 import marshal
 import stat
 import time
+
+py26_or_later = sys.version_info >= (2,6)
 
 
 class zope_ftp_channel(ftp_channel):
@@ -105,13 +108,25 @@ class zope_ftp_channel(ftp_channel):
         return path
 
     # Overriden async_chat methods
-
-    def push(self, producer, send=1):
+    def push(self, data, send=1):
         # this is thread-safe when send is false
         # note, that strings are not wrapped in
         # producers by default
-        self.producer_fifo.push(producer)
-        if send: self.initiate_send()
+
+        # LP #418454
+        if py26_or_later:
+            # Python 2.6 or later
+            sabs = self.ac_out_buffer_size
+            if len(data) > sabs:
+                for i in xrange(0, len(data), sabs):
+                    self.producer_fifo.append(data[i:i+sabs])
+            else:
+                self.producer_fifo.append(data)
+        else:
+            # pre-Python 2.6
+            self.producer_fifo.push(data)
+        if send: 
+            self.initiate_send()
 
     push_with_producer=push